requirements.txt
httpx>=0.24.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-pptx>=0.6.21
matplotlib>=3.7.0
Pillow>=9.5.0
pydantic>=2.0.0
typer>=0.9.0
playwright>=1.40.0
pyyaml>=6.0
